+++ packages/firebase_ui_database/example/lib/main.dart
@@ -20,7 +20,7 @@ Future<void> main() async {
 }
 
 class App extends StatelessWidget {
-  const App({Key? key}) : super(key: key);
+  const App({super.key});
 
   @override
   Widget build(BuildContext context) {