+++ packages/firebase_ui_shared/lib/src/loading_button.dart
@@ -2,22 +2,19 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:firebase_ui_auth/firebase_ui_auth.dart';
+import 'package:firebase_ui_shared/firebase_ui_shared.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-
-import 'universal_button.dart';
 
 class _LoadingButtonContent extends StatelessWidget {
   final String label;
   final bool isLoading;
   final Color? color;
   const _LoadingButtonContent({
-    Key? key,
     required this.label,
     required this.isLoading,
     required this.color,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -46,25 +43,50 @@ class _LoadingButtonContent extends StatelessWidget {
   }
 }
 
+/// Button widget that uses [CupertinoButton] under [CupertinoApp] and
+/// [TextButton], [ElevatedButton] or [OutlinedButton] under [MaterialApp]
+/// which is also capable of displaying a loading indicator when [isLoading] is
+/// set to true.
 class LoadingButton extends StatelessWidget {
+  /// Indicates that a loading indicator should be displayed.
   final bool isLoading;
+
+  /// The text to display in the button.
   final String label;
-  final IconData? icon;
-  final Color? color;
+
+  /// The icon to display in the button under [MaterialApp].
+  final IconData? materialIcon;
+
+  /// The icon to display in the button under [CupertinoApp].
+  final IconData? cupertinoIcon;
+
+  /// The color of the button background under [MaterialApp].
+  final Color? materialColor;
+
+  /// The color of the button background under [CupertinoApp].
+  final Color? cupertinoColor;
+
+  /// The color of the button content.
   final Color? labelColor;
+
+  /// A callback that is called when the button is pressed.
   final VoidCallback onTap;
-  final ButtonVariant? variant;
+
+  /// The variant of the button. See [ButtonVariant] for more information.
+  final ButtonVariant variant;
 
   const LoadingButton({
-    Key? key,
+    super.key,
     required this.label,
     required this.onTap,
     this.isLoading = false,
-    this.icon,
-    this.color,
+    this.materialIcon,
+    this.cupertinoIcon,
+    this.materialColor,
+    this.cupertinoColor,
     this.labelColor,
     this.variant = ButtonVariant.outlined,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -84,8 +106,10 @@ class LoadingButton extends StatelessWidget {
     );
 
     return UniversalButton(
-      color: color,
-      icon: icon,
+      materialColor: materialColor,
+      cupertinoColor: cupertinoColor,
+      materialIcon: materialIcon,
+      cupertinoIcon: cupertinoIcon,
       contentColor: contentColor,
       onPressed: onTap,
       variant: variant,