+++ packages/firebase_ui_shared/lib/src/loading_indicator.dart
@@ -2,22 +2,32 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:firebase_ui_shared/firebase_ui_shared.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import './widgets/internal/platform_widget.dart';
-
+/// A loading indicator that uses [CupertinoActivityIndicator] under
+/// [CupertinoApp] and [CircularProgressIndicator] under [MaterialApp].
+///
+/// Centered by default.
 class LoadingIndicator extends PlatformWidget {
+  /// The size of the loading indicator.
   final double size;
+
+  /// The width of the loading indicator's border.
+  /// This is only used for [CircularProgressIndicator] under [MaterialApp].
   final double borderWidth;
+
+  /// The color of the loading indicator.
+  /// This is only used for [CircularProgressIndicator] under [MaterialApp].
   final Color? color;
 
   const LoadingIndicator({
-    Key? key,
+    super.key,
     required this.size,
     required this.borderWidth,
     this.color,
-  }) : super(key: key);
+  });
 
   @override
   Widget? buildWrapper(BuildContext context, Widget child) {