+++ packages/firebase_ui_shared/lib/src/universal_button.dart
@@ -2,12 +2,11 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:firebase_ui_shared/firebase_ui_shared.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import 'platform_widget.dart';
-
-/// {@template ui.auth.widgets.button_variant}
+/// {@template ui.shared.widgets.button_variant}
 /// An enumeration of the possible button variants.
 /// {@endtemplate}
 enum ButtonVariant {
@@ -21,32 +20,57 @@ enum ButtonVariant {
   outlined,
 }
 
+/// Button widget that uses [CupertinoButton] under [CupertinoApp] and
+/// [TextButton], [ElevatedButton] or [OutlinedButton] under [MaterialApp]
+/// depending on provided [variant].
 class UniversalButton extends PlatformWidget {
+  /// A callback that is called when the button is pressed.
   final VoidCallback? onPressed;
+
+  /// The text to display in the button.
+  /// If [child] is provided, this will be ignored.
   final String? text;
+
+  /// The child to display in the button.
   final Widget? child;
-  final IconData? icon;
+
+  /// The icon to display in the button under [MaterialApp].
+  final IconData? materialIcon;
+
+  /// The icon to display in the button under [CupertinoApp].
+  final IconData? cupertinoIcon;
+
+  /// Defines the order of the icon and the label.
+  /// Icon will be placed on the left if [TextDirection.ltr] and on the right
+  /// if [TextDirection.rtl].
   final TextDirection? direction;
-  final ButtonVariant? variant;
-  final Color? color;
+
+  /// The variant of the button.
+  /// If not provided, [ButtonVariant.filled] will be used.
+  final ButtonVariant variant;
+
+  /// The color of the button background under [MaterialApp].
+  final Color? materialColor;
+
+  /// The color of the button background under [CupertinoApp].
+  final Color? cupertinoColor;
+
+  /// The color of the button content.
   final Color? contentColor;
 
   const UniversalButton({
-    Key? key,
+    super.key,
     this.text,
     this.child,
     this.onPressed,
-    this.icon,
+    this.materialIcon,
+    this.cupertinoIcon,
     this.direction = TextDirection.ltr,
-    this.variant,
-    this.color,
+    this.variant = ButtonVariant.filled,
+    this.materialColor,
+    this.cupertinoColor,
     this.contentColor,
-  })  : assert(text != null || child != null),
-        super(key: key);
-
-  ButtonVariant get _variant {
-    return variant ?? ButtonVariant.filled;
-  }
+  }) : assert(text != null || child != null);
 
   @override
   Widget buildCupertino(BuildContext context) {
@@ -56,16 +80,16 @@ class UniversalButton extends PlatformWidget {
       textDirection: direction,
       mainAxisAlignment: MainAxisAlignment.center,
       children: [
-        if (icon != null) ...[
+        if (cupertinoIcon != null) ...[
           if (direction == TextDirection.rtl) const SizedBox(width: 8),
-          Icon(icon, size: 20, color: contentColor),
+          Icon(cupertinoIcon, size: 20, color: contentColor),
           if (direction == TextDirection.ltr) const SizedBox(width: 8),
         ],
         this.child ?? Text(text!),
       ],
     );
 
-    if (_variant == ButtonVariant.text || _variant == ButtonVariant.outlined) {
+    if (variant == ButtonVariant.text || variant == ButtonVariant.outlined) {
       button = CupertinoButton(
         padding: EdgeInsets.zero,
         onPressed: onPressed,
@@ -78,9 +102,9 @@ class UniversalButton extends PlatformWidget {
       );
     }
 
-    if (color != null) {
+    if (cupertinoColor != null) {
       return CupertinoTheme(
-        data: CupertinoTheme.of(context).copyWith(primaryColor: color),
+        data: CupertinoTheme.of(context).copyWith(primaryColor: cupertinoColor),
         child: button,
       );
     } else {
@@ -94,31 +118,31 @@ class UniversalButton extends PlatformWidget {
 
     ButtonStyle? style;
 
-    if (color != null) {
+    if (materialColor != null) {
       MaterialStateColor? foregroundColor;
       MaterialStateColor? backgroundColor;
 
       if (variant == ButtonVariant.text) {
-        foregroundColor = MaterialStateColor.resolveWith((_) => color!);
+        foregroundColor = MaterialStateColor.resolveWith((_) => materialColor!);
       } else {
         foregroundColor = MaterialStateColor.resolveWith((_) => contentColor!);
-        backgroundColor = MaterialStateColor.resolveWith((_) => color!);
+        backgroundColor = MaterialStateColor.resolveWith((_) => materialColor!);
       }
 
       style = ButtonStyle(
         foregroundColor: foregroundColor,
         backgroundColor: backgroundColor,
         overlayColor: MaterialStateColor.resolveWith(
-          (states) => color!.withAlpha(20),
+          (states) => materialColor!.withAlpha(20),
         ),
       );
     }
 
-    if (icon != null) {
-      switch (_variant) {
+    if (materialIcon != null) {
+      switch (variant) {
         case ButtonVariant.text:
           return TextButton.icon(
-            icon: Icon(icon, color: contentColor),
+            icon: Icon(materialIcon, color: contentColor),
             onPressed: onPressed,
             label: child,
             style: style,
@@ -126,20 +150,20 @@ class UniversalButton extends PlatformWidget {
         case ButtonVariant.filled:
           return ElevatedButton.icon(
             onPressed: onPressed,
-            icon: Icon(icon, color: contentColor),
+            icon: Icon(materialIcon, color: contentColor),
             label: child,
             style: style,
           );
         case ButtonVariant.outlined:
           return OutlinedButton.icon(
             onPressed: onPressed,
-            icon: Icon(icon, color: contentColor),
+            icon: Icon(materialIcon, color: contentColor),
             label: child,
             style: style,
           );
       }
     } else {
-      switch (_variant) {
+      switch (variant) {
         case ButtonVariant.text:
           return TextButton(
             onPressed: onPressed,