+++ packages/firebase_ui_shared/lib/src/universal_scaffold.dart
@@ -2,22 +2,24 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:firebase_ui_shared/firebase_ui_shared.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import 'platform_widget.dart';
-
+/// Utility Scaffold that uses [CupertinoPageScaffold] under [CupertinoApp] and
+/// [Scaffold] under [MaterialApp].
 class UniversalScaffold extends PlatformWidget {
+  /// See [CupertinoPageScaffold.child] and [Scaffold.body]
   final Widget body;
 
   /// See [Scaffold.resizeToAvoidBottomInset]
   final bool? resizeToAvoidBottomInset;
 
   const UniversalScaffold({
-    Key? key,
+    super.key,
     required this.body,
     this.resizeToAvoidBottomInset,
-  }) : super(key: key);
+  });
 
   @override
   Widget buildCupertino(BuildContext context) {